bug fixed : user can only set the endtime of a section
