+++ frontend/components/home/ExploreGenres.tsx
@@ -80,7 +80,6 @@ const ExploreGenres = () => {
         left={0}
         zIndex={-99}
         borderRadius={"0.5em"}
-        boxShadow={"lg"}
       >
         <Text fontSize="2xl" fontWeight="bold" mb={"1em"} ml={"0.7em"}>
           Explore Genres
